convert generate_por_fixtures binary to #[ignore] test, remove separate binary
