+++ core/indexer/src/test_utils.rs
@@ -818,3 +818,114 @@ pub mod bls_test {
         agg.to_public_key().to_bytes()
     }
 }
+
+#[cfg(test)]
+mod fixture_gen {
+    use anyhow::{Result, anyhow};
+    use kontor_crypto::{
+        FileLedger, PorSystem,
+        api::{self, Challenge},
+    };
+    use serde::Serialize;
+
+    use super::valid_seed_field;
+
+    #[derive(Debug, Serialize)]
+    struct PorProofFixtures {
+        invalid_proof_hex: String,
+        cross_block_agg_hex: String,
+        note: String,
+    }
+
+    fn prepare_test_file(content: &[u8], filename: &str) -> (api::PreparedFile, api::FileMetadata) {
+        let mut nonce = [0u8; 32];
+        for (i, b) in filename.bytes().enumerate().take(32) {
+            nonce[i] = b;
+        }
+        api::prepare_file(content, filename, &nonce).expect("Failed to prepare file")
+    }
+
+    fn generate_invalid_proof_hex() -> Result<String> {
+        let file2_content = b"Second file with different content";
+        let (prepared_file2, metadata2) = prepare_test_file(file2_content, "file2.txt");
+        let mut ledger = FileLedger::new();
+        ledger.add_file(&metadata2)?;
+        let challenge = Challenge::new(
+            metadata2,
+            20000u64,
+            100,
+            valid_seed_field(99).field,
+            "node_1".to_string(),
+        );
+        let system = PorSystem::new(&ledger);
+        let proof = system
+            .prove(vec![&prepared_file2], std::slice::from_ref(&challenge))
+            .map_err(|e| anyhow!("Failed to generate proof: {e}"))?;
+        let mut bytes = proof
+            .to_bytes()
+            .map_err(|e| anyhow!("Failed to serialize proof: {e}"))?;
+        if let Some(last) = bytes.last_mut() {
+            *last ^= 0x01;
+        }
+        Ok(hex::encode(bytes))
+    }
+
+    fn generate_cross_block_agg_hex() -> Result<String> {
+        let (prepared_a, metadata_a) =
+            prepare_test_file(b"Content of file A for cross-block", "cross_a.txt");
+        let (prepared_b, metadata_b) =
+            prepare_test_file(b"Content of file B for cross-block", "cross_b.txt");
+        let (_prepared_c, metadata_c) =
+            prepare_test_file(b"Content of file C - new agreement", "cross_c.txt");
+        let mut ledger = FileLedger::new();
+        ledger.add_file(&metadata_a)?;
+        ledger.add_file(&metadata_b)?;
+        ledger.add_file(&metadata_c)?;
+        let challenges = vec![
+            Challenge::new(
+                metadata_a,
+                40000u64,
+                100,
+                valid_seed_field(200).field,
+                "node_1".to_string(),
+            ),
+            Challenge::new(
+                metadata_b,
+                40000u64,
+                100,
+                valid_seed_field(201).field,
+                "node_1".to_string(),
+            ),
+        ];
+        let system = PorSystem::new(&ledger);
+        let proof = system
+            .prove(vec![&prepared_a, &prepared_b], &challenges)
+            .map_err(|e| anyhow!("Failed to generate aggregated proof: {e}"))?;
+        let bytes = proof
+            .to_bytes()
+            .map_err(|e| anyhow!("Failed to serialize proof: {e}"))?;
+        Ok(hex::encode(bytes))
+    }
+
+    /// Regenerate POR proof fixtures. Run with:
+    /// `cargo test -p indexer --release --lib generate_por_fixtures -- --ignored`
+    #[test]
+    #[ignore]
+    fn generate_por_fixtures() {
+        let fixtures = PorProofFixtures {
+            invalid_proof_hex: generate_invalid_proof_hex().expect("invalid proof"),
+            cross_block_agg_hex: generate_cross_block_agg_hex().expect("cross block agg"),
+            note: "Generated by generate_por_fixtures".to_string(),
+        };
+        let output_path = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"))
+            .join("tests")
+            .join("fixtures")
+            .join("por_proof_fixtures.json");
+        if let Some(parent) = output_path.parent() {
+            std::fs::create_dir_all(parent).expect("create fixtures dir");
+        }
+        let json = serde_json::to_string_pretty(&fixtures).expect("serialize");
+        std::fs::write(&output_path, json).expect("write fixtures");
+        println!("Wrote fixtures to {}", output_path.display());
+    }
+}
